Fixed missing Division
Did this in two ways. First, I made it so team-opgg and tournament first check the division mapping. If its empty, respond eph with "No division". Secondly, i added a command that will query the Division Table and make any updates to the bot's information (since we store  and cache that table)